Webpack refactor complete - modular source code structure implemented. No functional adjustments.
